Make credentials not use Github secrets
The prototype is open source, we only use these as a layer of defence
gainst web crawlers and users accidentally stumbling upon it and
confusing it with the real thing.

This way Actions doesn't try to obfuscate the words `teaching` and
`vacancies` in the log output (which is annoying). And if you happen
to be reading this (yes, you), you can visit the prototype and enjoy
its beauty.